# server/openapi.json
{
  "openapi": "3.1.0",
  "servers": [
    {
      "url": "/api"
    }
  ],
  "info": {
    "title": "LocalEyes API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/reports/{id}/comments": {
      "post": {
        "summary": "Add a discussion comment to a report",
        "tags": [
          "Reports"
        ],
        "operationId": "createComment",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "description": "Report Id"
            },
            "required": true,
            "description": "Report Id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed (e.g. staff commenting outside their department)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Report not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Business rule violated (e.g. report is locked)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/reports/{id}/status": {
      "post": {
        "summary": "Change report status",
        "tags": [
          "Reports"
        ],
        "operationId": "updateReportStatus",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "description": "Report Id"
            },
            "required": true,
            "description": "Report Id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Status updated and status note created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusNoteResponse"
                }
              }
            }
          },
          "403": {
            "description": "Staff accessing a report outside their department",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Report not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Business rule violated (e.g. terminal status, same status)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateCommentRequest": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          }
        },
        "required": [
          "body"
        ]
      },
      "UpdateStatusRequest": {
        "type": "object",
        "properties": {
          "newStatus": {
            "type": "string",
            "enum": [
              "submitted",
              "acknowledged",
              "in_progress",
              "resolved",
              "closed",
              "rejected",
              "withdrawn"
            ]
          },
          "body": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          }
        },
        "required": [
          "newStatus",
          "body"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_FAILED",
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "DOMAIN_RULE",
                  "RATE_LIMITED",
                  "INTERNAL_ERROR",
                  "ROUTE_NOT_FOUND"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "CommentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "type": {
            "type": "string",
            "enum": [
              "discussion",
              "status_note"
            ]
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "newStatus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "authorName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "authorRole": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "citizen",
                  "staff",
                  "admin"
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "authorAvatarUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "isMine": {
            "type": "boolean"
          },
          "isHidden": {
            "type": "boolean"
          },
          "isEdited": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "body",
          "newStatus",
          "authorName",
          "authorRole",
          "authorAvatarUrl",
          "isMine",
          "isHidden",
          "isEdited",
          "createdAt"
        ],
        "additionalProperties": false
      },
      "StatusNoteResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "type": {
            "type": "string",
            "enum": [
              "discussion",
              "status_note"
            ]
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "newStatus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "authorName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "authorRole": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "citizen",
                  "staff",
                  "admin"
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "isMine": {
            "type": "boolean"
          },
          "isHidden": {
            "type": "boolean"
          },
          "isEdited": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "body",
          "newStatus",
          "authorName",
          "authorRole",
          "isMine",
          "isHidden",
          "isEdited",
          "createdAt"
        ],
        "additionalProperties": false
      },
...
    }
  }
}